matlab guy so set to keep up with binuclear Confrim_no returns 0 now. Acknowlege sends mssg back to matlab

+++ binoclean/stimuli.h
@@ -409,7 +409,7 @@ float *rcvals;
   int stimmode,ncalc,noverflow; 
   int fastextras,driveadapter;
   float fixpos[2];
-  int softoff[6];
+  float softoff[6];
   int hemisphere,electrode;
   double minval, fastincr;
   int nfast;
